requirements.txt
pandas>=2.0.0
matplotlib>=3.5.0
matplotlib-inline>=0.1.3
numpy>=1.24.0
Pillow>=9.0.0
tushare>=1.2.89
mplfinance>=0.12.9b7
typed-ast>=1.5.4
typer>=0.4.0
typing_extensions>=4.5.0
scikit-learn>=1.0.0
scipy>=1.10.0
tiktoken>=0.4.0
openai>=0.27.0
requests>=2.28.0
requests-oauthlib>=1.3.0
zhipuai>=2.0.1
dashscope>=1.14.0
gradio>=4.0.0
